add decoding.rs, hashing.rs, and data_provider.rs
