+++ wasm-host/Cargo.toml
@@ -4,11 +4,11 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0"
 clap = { version = "4.4", features = ["derive"] }
-serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 wasmedge-sdk = { version = "0.14.0", features = ["standalone"] }
 log = "0.4"
 env_logger = "0.11"
 sha2 = "0.11.0-pre.5"
+xrpl-rust = { version = "0.5.0", default-features = false, features = ["core"] }
+hex = "0.4.3"
