+++ wasm-host/src/host_functions.rs
@@ -1,263 +1,11 @@
 #![allow(dead_code)]
 
-use crate::mock_data::{DataSource, Hash256, Keylet, MockData};
-use sha2::{Digest, Sha512};
+use crate::data_provider::{unpack_locator, DataProvider, HostError};
+use crate::decoding::AccountId;
+use crate::hashing::{index_hash, sha512_half, LedgerNameSpace};
+use crate::mock_data::{DataSource, Keylet};
 use wasmedge_sdk::error::{CoreError, CoreExecutionError};
 use wasmedge_sdk::{CallingFrame, Instance, WasmValue};
-
-const LOCATOR_BUFFER_SIZE: usize = 64;
-const NUM_SLOTS: usize = 256;
-
-type AccountId = Vec<u8>; //TODO size
-
-pub enum HostError {
-    InternalError = -1,
-    FieldNotFound = -2,
-    BufferTooSmall = -3,
-    NoArray = -4,
-    NotLeafField = -5,
-    LocatorMalformed = -6,
-    SlotOutRange = -7,
-    SlotsFull = -8,
-    InvalidSlot = -9,
-    LedgerObjNotFound = -10,
-}
-
-pub struct LocatorUnpacker {
-    buffer: Vec<u8>,
-    cur_buffer_index: usize,
-}
-
-impl LocatorUnpacker {
-    pub fn from_bytes(buffer: Vec<u8>) -> Option<LocatorUnpacker> {
-        let packed_bytes: usize = buffer.len();
-        if packed_bytes > LOCATOR_BUFFER_SIZE || packed_bytes == 0 || packed_bytes % 4 != 0 {
-            None
-        } else {
-            Some(LocatorUnpacker {
-                buffer,
-                cur_buffer_index: 0,
-            })
-        }
-    }
-
-    pub fn unpack(&mut self) -> Option<i32> {
-        if self.cur_buffer_index + 4 > self.buffer.len() {
-            return None;
-        }
-        let mut bytes: [u8; 4] = [0u8; 4];
-        bytes.copy_from_slice(&self.buffer[self.cur_buffer_index..self.cur_buffer_index + 4]);
-        self.cur_buffer_index += 4;
-        Some(i32::from_le_bytes(bytes))
-    }
-}
-
-pub fn unpack_locator(buffer: Vec<u8>) -> Result<Vec<i32>, HostError> {
-    let mut unpacker = LocatorUnpacker::from_bytes(buffer).ok_or(HostError::LocatorMalformed)?;
-
-    let mut result = vec![];
-    while let Some(slot) = unpacker.unpack() {
-        result.push(slot);
-    }
-    Ok(result)
-}
-
-pub struct DataProvider {
-    data_source: MockData,
-    next_slot: usize,
-    slots: [Keylet; NUM_SLOTS],
-}
-
-impl DataProvider {
-    pub fn new(data_source: MockData) -> Self {
-        let slots: [Hash256; 256] = core::array::from_fn(|_| Hash256::default());
-        Self {
-            data_source,
-            next_slot: 1,
-            slots,
-        }
-    }
-
-    pub fn slot_set(&mut self, keylet: Keylet, mut slot: usize) -> i32 {
-        if slot == 0 {
-            if self.next_slot >= NUM_SLOTS {
-                return HostError::SlotsFull as i32;
-            }
-            slot = self.next_slot;
-            self.next_slot += 1;
-        } else if slot >= NUM_SLOTS {
-            return HostError::InvalidSlot as i32;
-        }
-
-        if self.data_source.obj_exist(&keylet) {
-            self.slots[slot] = keylet;
-            slot as i32
-        } else {
-            HostError::LedgerObjNotFound as i32
-        }
-    }
-
-    pub fn slot_get(&self, slot: usize) -> Option<&Keylet> {
-        if slot == 0 || slot >= self.next_slot {
-            None
-        } else {
-            Some(&self.slots[slot])
-        }
-    }
-
-    pub fn get_field_value(
-        &self,
-        source: DataSource,
-        idx_fields: Vec<i32>,
-        buf_cap: usize,
-    ) -> (i32, Vec<u8>) {
-        let field_result = self.data_source.get_field_value(source, idx_fields);
-        Self::fill_buf(field_result, buf_cap)
-    }
-
-    pub fn get_array_len(&self, source: DataSource, idx_fields: Vec<i32>) -> i32 {
-        match self.data_source.get_array_len(source, idx_fields) {
-            None => HostError::NoArray as i32,
-            Some(len) => len as i32,
-        }
-    }
-
-    pub fn get_ledger_sqn(&self, buf_cap: usize) -> (i32, Vec<u8>) {
-        let field_result = self.data_source.get_ledger_sqn();
-        Self::fill_buf(field_result, buf_cap)
-    }
-
-    pub fn get_parent_ledger_time(&self, buf_cap: usize) -> (i32, Vec<u8>) {
-        let field_result = self.data_source.get_parent_ledger_time();
-        Self::fill_buf(field_result, buf_cap)
-    }
-
-    pub fn get_parent_ledger_hash(&self, buf_cap: usize) -> (i32, Vec<u8>) {
-        let field_result = self.data_source.get_parent_ledger_hash();
-        Self::fill_buf(field_result, buf_cap)
-    }
-
-    pub fn set_current_ledger_obj_data(&mut self, data: Vec<u8>) {
-        self.data_source.set_current_ledger_obj_data(data);
-    }
-
-    pub fn fill_buf(field_result: Option<&serde_json::Value>, buf_cap: usize) -> (i32, Vec<u8>) {
-        let mut buf = vec![0u8;buf_cap];
-        match field_result {
-            Some(value) => {
-                match value {
-                    serde_json::Value::Number(n) => {
-                        if n.is_i64() {
-                            let num = n.as_i64().unwrap();
-                            if buf_cap == 4 {
-                                // Safe cast to i32
-                                if num > i32::MAX as i64 || num < i32::MIN as i64 {
-                                    return (HostError::BufferTooSmall as i32, buf);
-                                }
-                                let bytes = (num as i32).to_le_bytes();
-                                buf[..4].copy_from_slice(&bytes);
-                                (4, buf)
-                            } else {
-                                let bytes = num.to_le_bytes();
-                                if bytes.len() > buf_cap {
-                                    return (HostError::BufferTooSmall as i32, buf);
-                                }
-                                buf[..bytes.len()].copy_from_slice(&bytes);
-                                (bytes.len() as i32, buf)
-                            }
-                        } else if n.is_u64() {
-                            let num = n.as_u64().unwrap();
-                            if buf_cap == 4 {
-                                // Safe cast to u32
-                                if num > u32::MAX as u64 {
-                                    return (HostError::BufferTooSmall as i32, buf);
-                                }
-                                let bytes = (num as u32).to_le_bytes();
-                                buf[..4].copy_from_slice(&bytes);
-                                (4, buf)
-                            } else {
-                                let bytes = num.to_le_bytes();
-                                if bytes.len() > buf_cap {
-                                    return (HostError::BufferTooSmall as i32, buf);
-                                }
-                                buf[..bytes.len()].copy_from_slice(&bytes);
-                                (bytes.len() as i32, buf)
-                            }
-                        } else {
-                            return (HostError::InternalError as i32, buf);
-                        }
-                    }
-                    serde_json::Value::String(s) => {
-                        let bytes = s.as_bytes();
-                        if bytes.len() > buf_cap {
-                            return (HostError::BufferTooSmall as i32, buf);
-                        }
-                        buf[..bytes.len()].copy_from_slice(bytes);
-                        (bytes.len() as i32, buf)
-                    }
-                    _ => (HostError::InternalError as i32, buf),
-                }
-            }
-            None => (HostError::FieldNotFound as i32, buf),
-        }
-    }
-}
-
-#[repr(u16)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LedgerNameSpace {
-    Account = b'a' as u16,
-    DirNode = b'd' as u16,
-    TrustLine = b'r' as u16,
-    Offer = b'o' as u16,
-    OwnerDir = b'O' as u16,
-    BookDir = b'B' as u16,
-    SkipList = b's' as u16,
-    Escrow = b'u' as u16,
-    Amendments = b'f' as u16,
-    FeeSettings = b'e' as u16,
-    Ticket = b'T' as u16,
-    SignerList = b'S' as u16,
-    XrpPaymentChannel = b'x' as u16,
-    Check = b'C' as u16,
-    DepositPreauth = b'p' as u16,
-    DepositPreauthCredentials = b'P' as u16,
-    NegativeUnl = b'N' as u16,
-    NftokenOffer = b'q' as u16,
-    NftokenBuyOffers = b'h' as u16,
-    NftokenSellOffers = b'i' as u16,
-    Amm = b'A' as u16,
-    Bridge = b'H' as u16,
-    XchainClaimId = b'Q' as u16,
-    XchainCreateAccountClaimId = b'K' as u16,
-    Did = b'I' as u16,
-    Oracle = b'R' as u16,
-    MptokenIssuance = b'~' as u16,
-    Mptoken = b't' as u16,
-    Credential = b'D' as u16,
-    PermissionedDomain = b'm' as u16,
-
-    #[deprecated]
-    Contract = b'c' as u16,
-    #[deprecated]
-    Generator = b'g' as u16,
-    #[deprecated]
-    Nickname = b'n' as u16,
-}
-
-pub fn sha512_half(data: &[u8]) -> Hash256 {
-    let mut hasher = Sha512::new();
-    hasher.update(&data);
-    let result = hasher.finalize();
-    result[..32].to_vec()
-}
-
-pub fn index_hash(space: LedgerNameSpace, args: &[u8]) -> Hash256 {
-    let mut data = Vec::with_capacity(2 + args.len());
-    data.extend_from_slice(&(space as u16).to_le_bytes());
-    data.extend_from_slice(args);
-    sha512_half(&data)
-}
 
 fn get_data(
     in_buf_ptr: i32,
